+++ src/components/ComboBoxButton/ComboBoxButton.component.tsx
@@ -37,7 +37,7 @@ const defaultComboBoxButtonProps: ComboBoxButtonDefaultProps = {
 /**
  * A single clickable option in a ComboBox
  *
- * @version 0.0.3
+ * @version 0.0.4
  *
  * @param {ComboBoxButtonProps & TestIdProps} props - ComboBoxButton component props
  * @returns {JSX.Element}