+++ src/components/Input/ComboBoxInput/index.ts
@@ -1,0 +1,2 @@
+export * from "./ComboBoxInput.component";
+export type * from "./ComboBoxInput.types";